correct a bug of inheritance
isPawnDoubleSquareDisplacement was not able to recognize that a pawn is an instance of Pawn.
that because when we init the pieces of the Board, we use PieceJsonifier.
That class instanciate a new Piecefactory for every pieces to create.
I do not know realy why but the inistiate Piece was not correct,
so en passant capture eligibility was not correct.

I can not reproduce the case in the test...